Switched to Pipenv package manager and Added 'docs' with Sphinx
