--- README.md
@@ -1,2 +1,0 @@
-# handtracking
-Handtracking in Python using OpenCV and MediaPipe.
