+++ README.rst
@@ -1,0 +1,4 @@
+# Hand Tracking
+This is a Hand Tracking program made using OpenCV and MediaPipe.
+
+I made this following a YouTube video course, [Advanced Computer Vision with Python](https://youtu.be/01sAkU_NvOY) as part of my Seeing with Sound initiative.
