+++ docs/source/index.rst
@@ -1,0 +1,20 @@
+.. Hand Tracking documentation master file, created by
+   sphinx-quickstart on Thu Aug 12 11:40:37 2021.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to Hand Tracking's documentation!
+=========================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
