--- requirements.txt
@@ -1,2 +1,0 @@
-opencv-python~=4.5.3.56
-mediapipe~=0.8.6.2
